2023day05part1

+++ 2023/day05/part1.cpp
@@ -8,7 +8,7 @@ using namespace std;
 
 struct Range{
   
-  int beg, end, offset;
+  long beg, end, offset;
 
 };
 
@@ -16,7 +16,7 @@ struct Rule{
   
   vector<Range> ranges;
 
-  void apply(int & val){
+  void apply(long & val){
     
     for(Range & range : ranges)
       if(range.beg <= val && range.end >= val){
@@ -39,16 +39,16 @@ void read_rule(string & line, vector<Rule> & rules, int & rule_ctr){
 
   stringstream line_stream(line);
 
-  vector<int> nums;
+  vector<long> nums;
 
   while(getline(line_stream, field, ' '))
-    nums.push_back(stoi(field));
+    nums.push_back(stol(field));
 
   rules.back().ranges.push_back(Range{nums[1], nums[1] + nums[2] - 1, nums[0] - nums[1]});
 
 }
 
-void read_seeds(string & line, vector<int> & seeds){
+void read_seeds(string & line, vector<long> & seeds){
   
   string field;
 
@@ -56,7 +56,7 @@ void read_seeds(string & line, vector<int> & seeds){
 
   while(getline(line_stream, field, ' '))
     if(isdigit(field[0]))
-      seeds.push_back(stoi(field));
+      seeds.push_back(stol(field));
 
 }
 
@@ -66,13 +66,13 @@ int main(){
 
   ifstream in_file("input.txt");
   
-  vector<int> seeds;
+  vector<long> seeds;
   vector<Rule> rules;
 
   int rule_ctr = 0;
 
   while(getline(in_file, line)){
-
+     
     if(line.find("seeds:") != string::npos)
       read_seeds(line, seeds);
 
@@ -84,7 +84,7 @@ int main(){
 
   }
 
-  for(int & seed : seeds)
+  for(long & seed : seeds)
     for(Rule & rule : rules)
       rule.apply(seed);
